part2 solved by walking across the pipe
We find all the tiles that are part of the main pipe, iterate over
all tiles in the grid and whenever we cross the main pipe, count the
tiles we pass while inside. After passing the pipe again we are outside
and no longer count the tiles we pass over.

This is less involved than the previos solution that kept track of
the direction to the inside and walked inwards whenever possible.